Do not automatically set LIDORT thermal only mode as true.
The thermal only mode of LIDORT turns off scattering computations.
This speeds up the computation but causes problems when comparing
to VLIDORT or 2stream. In the case of 2stream it only provides
multiple scattering computations and therefore we are comparing just
the thermal scattering computation when comparing with it in the unit
tests.