Check point progress on converting code to use OpticalProperties class.
AtmosphereStandard and RtAtmosphere are converted to remove intermediate
variable logic and instead use OpticalProperty classes.

The code compiles, but unit tests still needs some work.

Some work was done on fleshing out modifications for the LSI to replace
the removed intermediate variable, but it is not finished.

PCA RT has been commented out till it can be reworked.