Switch Level1bSampleCoefficient so that in addition to spectral coefficients it also provides the variable values for evaluating the polynomial. In the past it was assumed that variable values used were a zero or one based array of indexes. This change makes the meaning of this polynomial more generic so it can still be retrieved but allow more flexibiltiy in its definition.
DispersionPolynomial has similarly been updated so it requires the
spectral variable values to be passed instead of the old is_one_based
flag. Additional cleanup was done to the DispersionPolynomial interface.
The Python configuration warns if the old is_one_based argument is used
to the IlsInstrument creator instead of variable values.

Also sped up the computation time of sample_grid in Level1bSampleCoefficient
by modifying to match the way it is computed in DispersionPolynomial.