Update Python creators to remove use of retrieval flag built into the StateVector. Instead use StateMappingAtIndexes where appropriate.
The primary retrieval input to many of the creators has been renamed
from 'value' to something more meaningful to the creator. The use of
'value' was a remnant of having a mixin that derived flags from the
retrieval coefficients. As a consequence configurations will need to be
updated to the new names.

The covariance creator still uses a retrieval flag to construct itself.
But instead it looks for a StateMappingAtIndexes mapping to be defined.
This could be a problem if a StateMappingAtIndexes is used inside a
StateMappingComposite mapping.