Updating ILS to account for units.
There are three objects that may have inconsistent types. The Instrument Sample Grid, the High Resolution Grid, and the High Res Extension (half-width). This commit updates ILS application so we convert the high resolution grid and the high res extension to the same units as the sample grid, if required. Also, we reverse the SpectralDomain (and corresponding reverse the SpectralRange), if needed in order to preserve ascending order in the SpectralDomain.